refactor(*) : Fix codesmells
2차 시도...
