+++ frontend/src/features/repository/RepositoryHeader.css
@@ -1,12 +1,8 @@
-.header {
-
-}
-
 #custom-switch {
     height: 30px !important;
     width: 60px !important;
 }
 
 .helper {
-    margin-top: 7.5px;
+    margin-top: 8px;
 }
