+++ frontend/src/features/route/PlaceDetail.tsx
@@ -129,7 +129,7 @@ export default function PlaceDetail() {
     }
 
     function onCheck(event : React.ChangeEvent<HTMLInputElement>) {
-        const id = parseInt(event.target.name) as number;
+        const id = parseInt(event.target.name);
         const temp = { ...checked };
         temp[id] = !checked[id];
         setChecked(temp);
